add middleware to purchase apis

+++ routes/api.php
@@ -173,12 +173,12 @@ Route::prefix('v1')->group(function () {
                 Route::post('/checkout', [PosController::class, 'checkoutCart'])->middleware('permissions:Create Transaction');
             });
             //purchase
-            Route::prefix('purchases')->group(function(){
-              Route::get('/', [PurchaseController::class, 'index']);
-              Route::get('/{id}',[PurchaseController::class, 'show']);
-              Route::post('/',[PurchaseController::class,'store']);
-              Route::patch('/{id}',[PurchaseController::class, 'update']);
-              Route::delete('/{id}',[PurchaseController::class,'destroy']);
+            Route::prefix('purchases')->middleware('modules:Purchases')->group(function(){
+              Route::get('/', [PurchaseController::class, 'index'])->middleware('permissions:View');
+              Route::get('/{id}',[PurchaseController::class, 'show'])->middleware('permissions:View');
+              Route::post('/',[PurchaseController::class,'store'])->middleware('permissions:Create');
+              Route::patch('/{id}',[PurchaseController::class, 'update'])->middleware('permissions:Edit');
+              Route::delete('/{id}',[PurchaseController::class,'destroy'])->middleware('permissions:Delete');
 
             });
         });
